+++ src/lib.rs
@@ -94,6 +94,7 @@ use webserver::Database;
 pub const TEMPLATES_DIR: &str = "sqlpage/templates/";
 pub const MIGRATIONS_DIR: &str = "migrations";
 pub const ON_CONNECT_FILE: &str = "on_connect.sql";
+pub const ON_RESET_FILE: &str = "on_reset.sql";
 
 pub struct AppState {
     pub db: Database,